# [TraceQL] Reduce work performed for structural operators due to a bug