**What this PR does**:
Prevents additional columns from being added as iterators when doing structural queries. For the query `{ resource.service.name = "loki-querier" } >> { resource.service.name = "loki-querier" }`, the following iterators were generated.

Before:
```
rebatchIterator: 
	JoinIterator: 0: traceCollector()	
		LeftJoinIterator: 1: batchCollector(true, 0)
		required: 
			LeftJoinIterator: 3: spanCollector(0)
			required: 
				SyncIterator: rs.list.element.ss.list.element.Spans.list.element.Kind : InstrumentedPredicate{2, nil}
			optional: 
				LeftJoinIterator: 4: attributeCollector{}
				required: 
					SyncIterator: rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.Key : InstrumentedPredicate{0, StringInPredicate{intrinsic(11), }}
				optional: 
					SyncIterator: rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.Value : InstrumentedPredicate{8, nil}
					SyncIterator: rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.ValueInt : InstrumentedPredicate{8, nil}
					SyncIterator: rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.ValueDouble : InstrumentedPredicate{8, nil}
					SyncIterator: rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.ValueBool : InstrumentedPredicate{8, nil}
				SyncIterator: rs.list.element.ss.list.element.Spans.list.element.NestedSetLeft : InstrumentedPredicate{2, nil}
				SyncIterator: rs.list.element.ss.list.element.Spans.list.element.NestedSetRight : InstrumentedPredicate{2, nil}
		optional: 
			SyncIterator: rs.list.element.Resource.ServiceName : InstrumentedPredicate{0, StringInPredicate{loki-querier, }}
			SyncIterator: rs.list.element.Resource.ServiceName : InstrumentedPredicate{0, StringInPredicate{loki-querier, }}
		SyncIterator: StartTimeUnixNano : InstrumentedPredicate{0, IntBetweenPredicate{0,1001000000000}}
		SyncIterator: EndTimeUnixNano : InstrumentedPredicate{0, IntBetweenPredicate{1000000000000,9223372036854775807}})
```

After:
```
rebatchIterator: 
	JoinIterator: 0: traceCollector()	
		LeftJoinIterator: 1: batchCollector(true, 0)
		required: 
			LeftJoinIterator: 3: spanCollector(0)
			required: 
				SyncIterator: rs.list.element.ss.list.element.Spans.list.element.Kind : InstrumentedPredicate{2, nil}
			optional: 
				SyncIterator: rs.list.element.ss.list.element.Spans.list.element.NestedSetLeft : InstrumentedPredicate{2, nil}
				SyncIterator: rs.list.element.ss.list.element.Spans.list.element.NestedSetRight : InstrumentedPredicate{2, nil}
		optional: 
			SyncIterator: rs.list.element.Resource.ServiceName : InstrumentedPredicate{0, StringInPredicate{loki-querier, }}
			SyncIterator: rs.list.element.Resource.ServiceName : InstrumentedPredicate{0, StringInPredicate{loki-querier, }}
		SyncIterator: StartTimeUnixNano : InstrumentedPredicate{0, IntBetweenPredicate{0,1001000000000}}
		SyncIterator: EndTimeUnixNano : InstrumentedPredicate{0, IntBetweenPredicate{1000000000000,9223372036854775807}})
```

Also finds a minor performance improvement for situations where a span has a nil intrinsic. This happens regularly for queries like `{ status = error }` where most entries in the StatusCode column are nil.

**Checklist**
- [ ] Tests updated
- [ ] Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`