@joe-elliott 2023-10-30
Initial benchmarks. I'm suspicious off the +1/2% and will re-run.

```
name                                          old time/op    new time/op    delta
BackendBlockTraceQL/spanAttNameNoMatch-8        6.20ms ± 3%    6.21ms ± 1%     ~     (p=0.310 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicMatch-8     42.9ms ± 1%    43.7ms ± 1%   +1.78%  (p=0.008 n=5+5)
BackendBlockTraceQL/statusError-8                122ms ± 1%     125ms ± 2%   +2.77%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttNameNoMatch-8    6.12ms ± 1%    6.23ms ± 2%   +1.90%  (p=0.032 n=5+5)
BackendBlockTraceQL/mixedNameNoMatch-8           407ms ± 2%     487ms ±20%     ~     (p=0.056 n=5+5)
BackendBlockTraceQL/desc-8                       2.09s ± 6%     1.63s ± 6%  -22.12%  (p=0.008 n=5+5)

name                                          old speed      new speed      delta
BackendBlockTraceQL/spanAttNameNoMatch-8       273MB/s ± 3%   273MB/s ± 1%     ~     (p=0.310 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicMatch-8    405MB/s ± 1%   398MB/s ± 1%   -1.75%  (p=0.008 n=5+5)
BackendBlockTraceQL/statusError-8              144MB/s ± 1%   140MB/s ± 2%   -2.69%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttNameNoMatch-8   190MB/s ± 1%   186MB/s ± 2%   -1.85%  (p=0.032 n=5+5)
BackendBlockTraceQL/mixedNameNoMatch-8        12.1MB/s ± 2%  10.3MB/s ±18%     ~     (p=0.056 n=5+5)
BackendBlockTraceQL/desc-8                    3.22MB/s ± 6%  1.35MB/s ± 6%  -57.92%  (p=0.008 n=5+5)

name                                          old MB_io/op   new MB_io/op   delta
BackendBlockTraceQL/spanAttNameNoMatch-8          1.69 ± 0%      1.69 ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicMatch-8       17.4 ± 0%      17.4 ± 0%     ~     (all equal)
BackendBlockTraceQL/statusError-8                 17.5 ± 0%      17.5 ± 0%     ~     (all equal)
BackendBlockTraceQL/resourceAttNameNoMatch-8      1.16 ± 0%      1.16 ± 0%     ~     (all equal)
BackendBlockTraceQL/mixedNameNoMatch-8            4.91 ± 0%      4.91 ± 0%     ~     (all equal)
BackendBlockTraceQL/desc-8                        6.72 ± 0%      2.20 ± 0%  -67.26%  (p=0.008 n=5+5)

name                                          old alloc/op   new alloc/op   delta
BackendBlockTraceQL/spanAttNameNoMatch-8        2.95MB ± 2%    2.93MB ± 3%     ~     (p=0.310 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicMatch-8     3.22MB ± 5%    3.31MB ± 6%     ~     (p=0.421 n=5+5)
BackendBlockTraceQL/statusError-8               14.1MB ± 2%    13.9MB ± 5%     ~     (p=0.222 n=5+5)
BackendBlockTraceQL/resourceAttNameNoMatch-8    2.95MB ± 3%    2.95MB ± 2%     ~     (p=0.690 n=5+5)
BackendBlockTraceQL/mixedNameNoMatch-8          7.48MB ±17%    6.78MB ±11%     ~     (p=0.310 n=5+5)
BackendBlockTraceQL/desc-8                      57.8MB ±84%   45.6MB ±107%     ~     (p=0.310 n=5+5)

name                                          old allocs/op  new allocs/op  delta
BackendBlockTraceQL/spanAttNameNoMatch-8         44.0k ± 0%     44.0k ± 0%     ~     (p=0.968 n=4+5)
BackendBlockTraceQL/spanAttIntrinsicMatch-8      45.4k ± 0%     45.4k ± 0%     ~     (p=0.587 n=5+5)
BackendBlockTraceQL/statusError-8                 142k ± 0%      142k ± 0%     ~     (p=0.381 n=5+5)
BackendBlockTraceQL/resourceAttNameNoMatch-8     44.1k ± 0%     44.1k ± 0%     ~     (p=1.000 n=5+5)
BackendBlockTraceQL/mixedNameNoMatch-8           44.7k ± 0%     44.7k ± 0%     ~     (p=0.452 n=5+5)
BackendBlockTraceQL/desc-8                       1.49M ± 6%     0.67M ±14%  -54.98%  (p=0.008 n=5+5)
```